Move files to folders
Move storage files and util files to their own folders for structure of
the project and convenience
