--- backend/storages.py
@@ -1,9 +1,0 @@
-from backend.users_storage import UsersStorage
-from backend.categories_storage import CategoriesStorage
-from backend.records_storage import RecordsStorage
-
-users = UsersStorage()
-
-categories = CategoriesStorage()
-
-records = RecordsStorage()
