+++ backend/storages/storages.py
@@ -1,0 +1,9 @@
+from backend.storages.users_storage import UsersStorage
+from backend.storages.categories_storage import CategoriesStorage
+from backend.storages.records_storage import RecordsStorage
+
+users = UsersStorage()
+
+categories = CategoriesStorage()
+
+records = RecordsStorage()
